requirements.txt
fastapi
uvicorn[standard]
requests
python-dateutil
pytz